Rename class to StoreExtensions, move to store package and update imports.
