+++ data/recommendedshows/implementation/src/commonMain/kotlin/com/thomaskioko/tvmaniac/data/recommendedshows/implementation/DefaultRecommendedShowsRepository.kt
@@ -2,7 +2,7 @@ package com.thomaskioko.tvmaniac.data.recommendedshows.implementation
 
 import com.thomaskioko.tvmaniac.core.base.model.AppCoroutineDispatchers
 import com.thomaskioko.tvmaniac.db.RecommendedShows
-import com.thomaskioko.tvmaniac.core.networkutil.mapToEither
+import com.thomaskioko.tvmaniac.core.store.mapToEither
 import com.thomaskioko.tvmaniac.core.networkutil.model.Either
 import com.thomaskioko.tvmaniac.core.networkutil.model.Failure
 import com.thomaskioko.tvmaniac.data.recommendedshows.api.RecommendedShowsParams
