+++ data/search/implementation/src/commonMain/kotlin/com/thomaskioko/tvmaniac/search/implementation/DefaultSearchRepository.kt
@@ -1,7 +1,7 @@
 package com.thomaskioko.tvmaniac.search.implementation
 
 import com.thomaskioko.tvmaniac.core.base.model.AppCoroutineDispatchers
-import com.thomaskioko.tvmaniac.core.networkutil.mapToEither
+import com.thomaskioko.tvmaniac.core.store.mapToEither
 import com.thomaskioko.tvmaniac.core.networkutil.model.Either
 import com.thomaskioko.tvmaniac.core.networkutil.model.Failure
 import com.thomaskioko.tvmaniac.search.api.SearchRepository
