+++ data/similar/implementation/src/commonMain/kotlin/com/thomaskioko/tvmaniac/similar/implementation/DefaultSimilarShowsRepository.kt
@@ -2,7 +2,7 @@ package com.thomaskioko.tvmaniac.similar.implementation
 
 import com.thomaskioko.tvmaniac.core.base.model.AppCoroutineDispatchers
 import com.thomaskioko.tvmaniac.db.SimilarShows
-import com.thomaskioko.tvmaniac.core.networkutil.mapToEither
+import com.thomaskioko.tvmaniac.core.store.mapToEither
 import com.thomaskioko.tvmaniac.core.networkutil.model.Either
 import com.thomaskioko.tvmaniac.core.networkutil.model.Failure
 import com.thomaskioko.tvmaniac.resourcemanager.api.RequestManagerRepository
